+++ src/tools/rerunWorkflow/inputSchema.ts
@@ -10,7 +10,7 @@ export const rerunWorkflowInputSchema = z.object({
   fromFailed: z
     .boolean()
     .describe(
-      'This should be true by default. If the user wants to rerun workflow from started, set this to false',
+      'This should be true by default. If the user wants to rerun workflow from start, set this to false',
     )
     .optional(),
   workflowURL: z